+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+
+#### v1.5
+
+- Adding new [report mode](https://github.com/j3ssie/Osmedeus/wiki/report) for easily view result from command line.
+- Change default path of config and result to `~/.osmedeus/`.
+- Change some routine in dirbrute module.
+- Fix a lot of bugs and refactoring a lot of things.
+
 #### v1.4
+
 - Adding new AssetFinding module powered mostly by [tomnomnom](https://github.com/tomnomnom).
 - Direct mode (specific module) now very powerful [Detail](https://github.com/j3ssie/Osmedeus/wiki/Direct-Mode).
 - Improve the API architecture, from now you can run multi targets without crash the routine.
@@ -7,8 +16,8 @@
 - Improve search and sort from the UI.
 - Fix a lot of bugs and refactoring a lot of things.
 
-
 #### v1.3
+
 - Support Continuous Scan.
 - Improve the routine.
 - Fix a lot of bugs.
@@ -20,8 +29,8 @@
 - Adding Dockerfile.
 - Support proxy via proxychains.
 
-
 #### v1.2
+
 - Fix a lot of bugs.
 - Improve the API architecture.
 - Adding JWT token for API.
@@ -30,12 +39,14 @@
 
 
 #### v1.1
+
 - Improve architecture.
 - React Web UI.
 - Fix a lot of bugs.
 
 
 #### v1.0
+
 - Collection of awesome tools againt the target. Details
 - Seperate workspaces to store all scan output and details logging.
 - REST API.